@@ -1,31 +1,34 @@
-import React from 'react';
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
 
 import { isSupported, visibility } from './visibility';
 
-export default React.createClass({
-    displayName: 'PageVisibility',
-    propTypes: {
-        onChange: React.PropTypes.func.isRequired,
-    },
+export default class PageVisibility extends Component {
     componentWillMount() {
         if (!isSupported || !visibility) {
             return;
         }
+
+        this.handleVisibilityChange = this.handleVisibilityChange.bind(this);
         this.isListening = true;
+
         document.addEventListener(visibility.event, this.handleVisibilityChange);
-    },
+    }
+
     componentWillUnmount() {
         if (!this.isListening) {
             return;
         }
         document.removeEventListener(visibility.event, this.handleVisibilityChange);
-    },
+    }
+
     handleVisibilityChange() {
         const { onChange } = this.props;
         const { hidden, state } = visibility;
 
         onChange(document[state], document[hidden]);
-    },
+    }
+
     render() {
         if (!this.props.children) {
             return null;
@@ -33,4 +36,10 @@ export default React.createClass({
         const child = React.Children.only(this.props.children);
         return React.cloneElement(child);
     }
-});
+}
+
+PageVisibility.displayName = 'PageVisibility';
+
+PageVisibility.propTypes = {
+    onChange: PropTypes.func.isRequired
+};
